create dev-server rollup config
